// src/enums/environment_map.ts
export enum EnvironmentMap0 {
  NX = '../../assets/textures/environment_maps/0/nx.jpg',
  NY = '../../assets/textures/environment_maps/0/ny.jpg',
  NZ = '../../assets/textures/environment_maps/0/nz.jpg',
  PX = '../../assets/textures/environment_maps/0/px.jpg',
  PY = '../../assets/textures/environment_maps/0/py.jpg',
  PZ = '../../assets/textures/environment_maps/0/pz.jpg',
}

export enum EnvironmentMap1 {
  NX = '../../assets/textures/environment_maps/1/nx.jpg',
  NY = '../../assets/textures/environment_maps/1/ny.jpg',
  NZ = '../../assets/textures/environment_maps/1/nz.jpg',
  PX = '../../assets/textures/environment_maps/1/px.jpg',
  PY = '../../assets/textures/environment_maps/1/py.jpg',
  PZ = '../../assets/textures/environment_maps/1/pz.jpg',
}

export enum EnvironmentMap2 {
  NX = '../../assets/textures/environment_maps/2/nx.png',
  NY = '../../assets/textures/environment_maps/2/ny.png',
  NZ = '../../assets/textures/environment_maps/2/nz.png',
  PX = '../../assets/textures/environment_maps/2/px.png',
  PY = '../../assets/textures/environment_maps/2/py.png',
  PZ = '../../assets/textures/environment_maps/2/pz.png',
}